rename package to webauthn
