+++ internal/webauthn/webauthn_test.go
@@ -1,22 +1,20 @@
-package passkey_test
+package webauthn_test
 
 import (
-	"fmt"
 	"testing"
 
-	"github.com/openauth/openauth/internal/passkey"
+	"github.com/openauth/openauth/internal/webauthn"
+	"github.com/stretchr/testify/require"
 )
 
 func TestParse(t *testing.T) {
-	c, err := passkey.Parse(&passkey.ParseRequest{
+	c, err := webauthn.Parse(&webauthn.ParseRequest{
 		RPID:              "localhost",
 		AttestationObject: "o2NmbXRkbm9uZWdhdHRTdG10oGhhdXRoRGF0YViYSZYN5YgOjGh0NBcPZHZgW4_krrmihjLHmVzzuoMdl2NdAAAAAPv8MAcVTk7MjAtuAgVX170AFFOOoV7GrKnKnccDm-8m0dTm_yQFpQECAyYgASFYILSxMv1cg3WN5GkouhyLJOXrIbBgSi9yAjI_QrC-IhAIIlggTDqhDZTjsPzx-dq2lkfu2AiieZuIpPUpQtOBYrX8gJw",
 	})
-	if err != nil {
-		panic(err)
-	}
+	require.NoError(t, err)
 
-	fmt.Println(c.Verify(&passkey.VerifyRequest{
+	require.NoError(t, c.Verify(&webauthn.VerifyRequest{
 		RPID:              "localhost",
 		Origin:            "http://localhost:3002",
 		ChallengeSHA256:   []byte{132, 143, 185, 51, 158, 98, 37, 12, 23, 156, 66, 204, 255, 170, 216, 93, 168, 10, 69, 31, 108, 79, 71, 89, 15, 138, 213, 219, 29, 51, 128, 200},
